Removing c89 flag.  Removing long int.  Making method static.

+++ t0/mocks/FakeTimeService.c
@@ -1,9 +1,9 @@
 /***
  * Excerpted from "Test-Driven Development for Embedded C",
  * published by The Pragmatic Bookshelf.
- * Copyrights apply to this code. It may not be used to create training material, 
+ * Copyrights apply to this code. It may not be used to create training material,
  * courses, books, articles, and the like. Contact us if you are in doubt.
- * We make no guarantees that this code is fit for any purpose. 
+ * We make no guarantees that this code is fit for any purpose.
  * Visit http://www.pragmaticprogrammer.com/titles/jgade for more book information.
 ***/
 /*- ------------------------------------------------------------------ -*/
@@ -26,7 +26,7 @@
 
 #include "FakeTimeService.h"
 
-static long int theMinute;
+static int theMinute;
 static int theDay;
 
 void TimeService_Create(void)
@@ -65,4 +65,3 @@ int TimeService_GetDay(void)
 {
     return theDay;
 }
-
